Simplify invoice print for admin
